Add files via upload
In this example I show  reading of an Adafruit ADS1015 module from a ESP32
The contribution is to replace the arduino wire.h library with the espresif "driver/i2c.h" library.   

I started from the Adafruit_ADS1X15.h library and the singleended sketch.

Basically I modified two functions from the Adafruit_ADS1X15.h library 
           -    void Adafruit_ADS1X15::writeRegister(uint8_t reg, uint16_t value)
           -     uint16_t Adafruit_ADS1X15::readRegister(uint8_t reg)

Replacing the calls to wire.h by the calls to the espresif library  according the programming guide. 
Remark the I2C initialization is done in the sketch set-up.  Which makes it easier to make changes such as the I2C speed.